renamed SAMPLE.jsx file
