--- src/Components/Sample.jsx
@@ -1,13 +1,0 @@
-import logo from "../logo.svg";
-import "../App.css";
-
-function Sample() {
-  return (
-    <>
-      <h1>Sample component</h1>
-      <img src={logo} className="App-logo" alt="logo" />
-    </>
-  );
-}
-
-export default Sample;
