+++ src/Components/Sample/StoryBookExampleComponent.jsx
@@ -1,0 +1,15 @@
+import React from "react";
+import logo from "../logo.svg";
+import "../App.css";
+
+
+function StoryBookExampleComponent() {
+  return (
+    <>
+      <h1>Sample component in community blog</h1>
+      <img src={logo} className="App-logo" alt="logo" />
+    </>
+  );
+}
+
+export default StoryBookExampleComponent;
